requirement.txt
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
psycopg2==2.9.6
Pillow==9.5.0